Trunk 1.1: new API - grp() without args - create empty group - fix god mode on actor attribute player type mismatch (string vs. number) - adjust inconsistent collision behaviour of jumping marbles, pearls with     rotors, tops, and killers:     - let balls jumps unharmed over rotors, killers     - let ball shatter or collide, if shielded, on jumps over tops - fl_water on flood behaviour: make a flood visible when it crosses an     existing idle water lake

+++ src/actors/Balls.cc
@@ -119,7 +119,7 @@ namespace enigma {
         // Shield, booze and invisibility can be activated in all states except DEAD
     
         if (state != DEAD) {
-            if (m.message == "_shield" && getAttr("owner") ==  m.value) {
+            if (m.message == "_shield" && getAttr("owner") ==  (int)m.value) {
                 m_shield_rest_time += SHIELD_TIME;
                 update_halo();
                 return true;
